requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.1
drf-yasg==1.21.4

